HistogramChartManager, Volume,BufferedVolume,HasHistogram: Move generating histogram to thier specific classes (Volume, BufferdVolume) and use the interface in a more sensible way.
