+++ src/main/kotlin/graphics/scenery/volumes/HistogramChartManager.kt
@@ -1,21 +1,18 @@
 package graphics.scenery.volumes
 
-import graphics.scenery.SceneryElement
 import org.jfree.chart.ChartPanel
 import org.jfree.chart.axis.LogarithmicAxis
 import org.jfree.chart.axis.NumberAxis
 import org.jfree.chart.plot.XYPlot
 import org.jfree.chart.renderer.xy.StandardXYBarPainter
 import org.jfree.chart.renderer.xy.XYBarRenderer
-import org.jfree.data.statistics.SimpleHistogramBin
 import org.jfree.data.statistics.SimpleHistogramDataset
 import java.awt.Color
 import javax.swing.JCheckBox
 import kotlin.math.abs
-import kotlin.math.max
 
 /**
- * Handles all histogram related things.
+ * Handles all histogram chart related things.
  *
  * @author Jan Tiemann <j.tiemann@hzdr.de>
  */
@@ -65,7 +62,7 @@ class HistogramChartManager(val tfPlot: XYPlot,
                 } else {
                     tfPlot.setDataset(1, volumeHistogramData)
                     tfPlot.setRangeAxis(1, histYAxis)
-                    populateHistogramBins( volumeHistogramData)
+                    generateHistogram( volumeHistogramData)
 
                     tfPlot.setDomainAxis(1, histXAxis)
 
@@ -75,95 +72,20 @@ class HistogramChartManager(val tfPlot: XYPlot,
         }
     }
 
-    private fun populateHistogramBins(volumeHistogramData: SimpleHistogramDataset) {
+    private fun generateHistogram(volumeHistogramData: SimpleHistogramDataset) {
         volumeHistogramData.removeAllBins()
-
-        val volume = tfContainer as? BufferedVolume
-        if (volume == null){
-            cpuGenerateHistogramBins(volumeHistogramData)
-            return
-        }
-
-        val volumeHistogram = VolumeHistogramComputeNode.generateHistogram(
-            volume,
-            volume.timepoints?.get(volume.currentTimepoint)!!.contents,
-            volume.getScene() ?: return
-        )
-        val histogram = volumeHistogram.fetchHistogram(volume.getScene()!!, volume.volumeManager.hub!!.get<graphics.scenery.backends.Renderer>(
-            SceneryElement.Renderer
-        )!!)
-
-        val displayRange = abs(tfContainer.maxDisplayRange - tfContainer.minDisplayRange)
-        val binSize = displayRange / volumeHistogram.numBins
-        val minDisplayRange = tfContainer.minDisplayRange.toDouble()
-
-        var max = 0
-        histogram.forEachIndexed { index, value ->
-            val bin = SimpleHistogramBin(
-                minDisplayRange + index * binSize,
-                minDisplayRange + (index + 1) * binSize,
-                true,
-                false
-            )
-            bin.itemCount = value
-            volumeHistogramData.addBin(bin)
-            max = max(max,value)
-        }
+        val max = (tfContainer as? HasHistogram)?.generateHistogram(volumeHistogramData) ?: return
 
         histYAxis.setRange(
             -0.1 ,
             max * (1.2)
         )
+
+        val displayRange = abs(tfContainer.maxDisplayRange - tfContainer.minDisplayRange)
         histXAxis.setRange(
             tfContainer.minDisplayRange - (axisExtensionFactor * displayRange),
             tfContainer.maxDisplayRange + (axisExtensionFactor * displayRange)
         )
     }
 
-    /**
-     *  Calculates the histogram on the CPU.
-     */
-    private fun cpuGenerateHistogramBins(volumeHistogramData: SimpleHistogramDataset) {
-        volumeHistogramData.removeAllBins()
-
-        // This generates a histogram over the whole volume ignoring the display range.
-        // We now need to select only the bins we care about.
-        val absoluteHistogram = (tfContainer as? HasHistogram)?.generateHistogram()
-        if (absoluteHistogram != null) {
-            val displayRange = abs(tfContainer.maxDisplayRange - tfContainer.minDisplayRange)
-
-            val absoluteBinSize = absoluteHistogram.max() / 1024.0
-            val minDisplayRange = tfContainer.minDisplayRange.toDouble()
-            val maxDisplayRange = tfContainer.maxDisplayRange.toDouble()
-
-            var max = 100
-            absoluteHistogram.forEachIndexed { index, longType ->
-                val startOfAbsoluteBin = index * absoluteBinSize
-                val endOfAbsoluteBin = (index+1) * absoluteBinSize
-                if (minDisplayRange <= startOfAbsoluteBin && endOfAbsoluteBin < maxDisplayRange) {
-
-                    val bin = SimpleHistogramBin(
-                        startOfAbsoluteBin,
-                        endOfAbsoluteBin,
-                        true,
-                        false
-                    )
-                    bin.itemCount = longType.get().toInt()
-                    max = max(bin.itemCount, max)
-                    volumeHistogramData.addBin(bin)
-                }
-            }
-
-            histYAxis.setRange(
-                -0.1 ,
-                max * (1.2)
-            )
-
-            histXAxis.setRange(
-                tfContainer.minDisplayRange - (axisExtensionFactor * displayRange),
-                tfContainer.maxDisplayRange + (axisExtensionFactor * displayRange)
-            )
-
-        }
-    }
 }
